bugfix for int2hex function

+++ src/utils.h
@@ -29,7 +29,7 @@ void str_trim_doubles(char *dst, char *src, char c);
 
 void str_replace_char(char *dst, char *src, char from, char to);
 
-void int2hex(char *dst, unsigned int input, size_t dim);
+void int2hex(char *dst, unsigned int input, int dim);
 
 unsigned int hex2uint(char *input);
